Cleaned and updated the tests
