+++ applications/StructuralMechanicsApplication/tests/test_StructuralMechanicsApplication.py
@@ -16,8 +16,9 @@ except ImportError as e:
     missing_application = re.search(r'''.*'KratosMultiphysics\.(.*)'.*''',
                                     '{0}'.format(e)).group(1)
 
-# Import the tests o test_classes to create the suits
-## SMALL 
+# Import the tests or test_classes to create the suits
+
+##### SELF-CONTAINED TESTS #####
 # CL tests
 from constitutive_law_test import TestConstitutiveLaw as TTestConstitutiveLaw
 # Simple patch tests
@@ -31,65 +32,11 @@ from test_patch_test_cr_beam import TestCrBeam2D2N as TTestCrBeam2D2N
 from test_patch_test_shells_stress import TestPatchTestShellsStressRec as TTestPatchTestShellsStressRec
 from test_patch_test_shells_orthotropic import TestPatchTestShellsOrthotropic as TTestPatchTestShellsOrthotropic
 # Test loading conditions
-from test_loading_conditions_point import TestLoadingConditionsPoint as TestLoadingConditionsPoint
-from test_loading_conditions_line import TestLoadingConditionsLine as TestLoadingConditionsLine
-from test_loading_conditions_surface import TestLoadingConditionsSurface as TestLoadingConditionsSurface
-# Basic moving mesh test
-from SmallTests import SimpleMeshMovingTest as TSimpleMeshMovingTest
-# Dynamic basic tests
-from SmallTests import DynamicBossakTests as TDynamicBossakTests
-from SmallTests import DynamicNewmarkTests as TDynamicNewmarkTests
-# Patch test Small Displacements
-from SmallTests import SDTwoDShearQuaPatchTest as TSDTwoDShearQuaPatchTest
-from SmallTests import SDTwoDShearTriPatchTest as TSDTwoDShearTriPatchTest
-from SmallTests import SDTwoDTensionQuaPatchTest as TSDTwoDTensionQuaPatchTest
-from SmallTests import SDTwoDTensionTriPatchTest as TSDTwoDTensionTriPatchTest
-from SmallTests import SDThreeDShearHexaPatchTest as TSDThreeDShearHexaPatchTest
-from SmallTests import SDThreeDShearTetraPatchTest as TSDThreeDShearTetraPatchTest
-from SmallTests import SDThreeDTensionHexaPatchTest as TSDThreeDTensionHexaPatchTest
-from SmallTests import SDThreeDTensionTetraPatchTest as TSDThreeDTensionTetraPatchTest
-# Patch test Total Lagrangian
-from SmallTests import TLTwoDShearQuaPatchTest as TTLTwoDShearQuaPatchTest
-from SmallTests import TLTwoDShearTriPatchTest as TTLTwoDShearTriPatchTest
-from SmallTests import TLTwoDTensionQuaPatchTest as TTLTwoDTensionQuaPatchTest
-from SmallTests import TLTwoDTensionTriPatchTest as TTLTwoDTensionTriPatchTest
-from SmallTests import TLThreeDShearHexaPatchTest as TTLThreeDShearHexaPatchTest
-from SmallTests import TLThreeDShearTetraPatchTest as TTLThreeDShearTetraPatchTest
-from SmallTests import TLThreeDTensionHexaPatchTest as TTLThreeDTensionHexaPatchTest
-from SmallTests import TLThreeDTensionTetraPatchTest as TTLThreeDTensionTetraPatchTest
-# Patch test Updated Lagrangian
-from SmallTests import ULTwoDShearQuaPatchTest as TULTwoDShearQuaPatchTest
-from SmallTests import ULTwoDShearTriPatchTest as TULTwoDShearTriPatchTest
-from SmallTests import ULTwoDTensionQuaPatchTest as TULTwoDTensionQuaPatchTest
-from SmallTests import ULTwoDTensionTriPatchTest as TULTwoDTensionTriPatchTest
-from SmallTests import ULThreeDShearHexaPatchTest as TULThreeDShearHexaPatchTest
-from SmallTests import ULThreeDShearTetraPatchTest as TULThreeDShearTetraPatchTest
-from SmallTests import ULThreeDTensionHexaPatchTest as TULThreeDTensionHexaPatchTest
-from SmallTests import ULThreeDTensionTetraPatchTest as TULThreeDTensionTetraPatchTest
-# SPRISM tests
-from SmallTests import SprismMembranePatchTests as TSprismMembranePatchTests
-from SmallTests import SprismBendingPatchTests as TSprismBendingPatchTests
-# Eigenvalues tests
-from SmallTests import EigenQ4Thick2x2PlateTests as TEigenQ4Thick2x2PlateTests
-from SmallTests import EigenTL3D8NCubeTests as TEigenTL3D8NCubeTests
-from SmallTests import Eigen3D3NThinCircleTests as TEigen3D3NThinCircleTests
-# Membrane tests
-from SmallTests import Fofi4PointTentnoCableTests as TFofi4PointTentnoCableTests
-from SmallTests import Fofi4PointTentCableTests as TFofi4PointTentCableTests
-from SmallTests import MembraneQ4PointLoadTests as TMembraneQ4PointLoadTests
-from SmallTests import MembraneQ4TrussPointLoadTests as TMembraneQ4TrussPointLoadTests
-# 2Node Element tests
-from SmallTests import Simple3D2NTrussTest as T3D2NTrussTest
-from SmallTests import Simple3D2NTrussLinearTest as T3D2NTrussLinearTest
-from SmallTests import Simple3D2NTrussDynamicTest as T3D2NTrussDynamicTest
-from SmallTests import Simple3D2NBeamCrTest as T3D2NBeamCrTest
-from SmallTests import Simple3D2NBeamCrLinearTest as T3D2NBeamCrLinearTest
-from SmallTests import Simple3D2NBeamCrDynamicTest as T3D2NBeamCrDynamicTest
-from SmallTests import Simple2D2NBeamCrTest as T2D2NBeamCrTest
-
+from test_loading_conditions_point import TestLoadingConditionsPoint as TTestLoadingConditionsPoint
+from test_loading_conditions_line import TestLoadingConditionsLine as TTestLoadingConditionsLine
+from test_loading_conditions_surface import TestLoadingConditionsSurface as TTestLoadingConditionsSurface
 # Multipoint constraint tests
 from test_multipoint_contstraints import TestMultipointConstraints as TTestMultipointConstraints
-
 # Nodal damping test
 from test_nodal_damping import NodalDampingTests as TNodalDampingTests
 # Spring damper element test
@@ -97,23 +44,76 @@ from test_spring_damper_element import SpringDamperElementTests as TSpringDamper
 # Harmonic analysis tests
 from test_harmonic_analysis import HarmonicAnalysisTests as THarmonicAnalysisTests
 
-## NIGHTLY TESTS
-# Shell test
+
+##### SMALL TESTS #####
+# Dynamic basic tests (leave these in the smallSuite to have the Exection script tested)
+from SmallTests import DynamicBossakTests as TDynamicBossakTests
+from SmallTests import DynamicNewmarkTests as TDynamicNewmarkTests
+
+
+##### NIGHTLY TESTS #####
+# Basic moving mesh test
+from NightlyTests import SimpleMeshMovingTest as TSimpleMeshMovingTest
+# Patch test Small Displacements
+from NightlyTests import SDTwoDShearQuaPatchTest as TSDTwoDShearQuaPatchTest
+from NightlyTests import SDTwoDShearTriPatchTest as TSDTwoDShearTriPatchTest
+from NightlyTests import SDTwoDTensionQuaPatchTest as TSDTwoDTensionQuaPatchTest
+from NightlyTests import SDTwoDTensionTriPatchTest as TSDTwoDTensionTriPatchTest
+from NightlyTests import SDThreeDShearHexaPatchTest as TSDThreeDShearHexaPatchTest
+from NightlyTests import SDThreeDShearTetraPatchTest as TSDThreeDShearTetraPatchTest
+from NightlyTests import SDThreeDTensionHexaPatchTest as TSDThreeDTensionHexaPatchTest
+from NightlyTests import SDThreeDTensionTetraPatchTest as TSDThreeDTensionTetraPatchTest
+# Patch test Total Lagrangian
+from NightlyTests import TLTwoDShearQuaPatchTest as TTLTwoDShearQuaPatchTest
+from NightlyTests import TLTwoDShearTriPatchTest as TTLTwoDShearTriPatchTest
+from NightlyTests import TLTwoDTensionQuaPatchTest as TTLTwoDTensionQuaPatchTest
+from NightlyTests import TLTwoDTensionTriPatchTest as TTLTwoDTensionTriPatchTest
+from NightlyTests import TLThreeDShearHexaPatchTest as TTLThreeDShearHexaPatchTest
+from NightlyTests import TLThreeDShearTetraPatchTest as TTLThreeDShearTetraPatchTest
+from NightlyTests import TLThreeDTensionHexaPatchTest as TTLThreeDTensionHexaPatchTest
+from NightlyTests import TLThreeDTensionTetraPatchTest as TTLThreeDTensionTetraPatchTest
+# Patch test Updated Lagrangian
+from NightlyTests import ULTwoDShearQuaPatchTest as TULTwoDShearQuaPatchTest
+from NightlyTests import ULTwoDShearTriPatchTest as TULTwoDShearTriPatchTest
+from NightlyTests import ULTwoDTensionQuaPatchTest as TULTwoDTensionQuaPatchTest
+from NightlyTests import ULTwoDTensionTriPatchTest as TULTwoDTensionTriPatchTest
+from NightlyTests import ULThreeDShearHexaPatchTest as TULThreeDShearHexaPatchTest
+from NightlyTests import ULThreeDShearTetraPatchTest as TULThreeDShearTetraPatchTest
+from NightlyTests import ULThreeDTensionHexaPatchTest as TULThreeDTensionHexaPatchTest
+from NightlyTests import ULThreeDTensionTetraPatchTest as TULThreeDTensionTetraPatchTest
+# SPRISM tests
+from NightlyTests import SprismMembranePatchTests as TSprismMembranePatchTests
+from NightlyTests import SprismBendingPatchTests as TSprismBendingPatchTests
+# Eigenvalues tests
+from NightlyTests import EigenQ4Thick2x2PlateTests as TEigenQ4Thick2x2PlateTests
+from NightlyTests import EigenTL3D8NCubeTests as TEigenTL3D8NCubeTests
+from NightlyTests import Eigen3D3NThinCircleTests as TEigen3D3NThinCircleTests
+# Membrane tests
+from NightlyTests import Fofi4PointTentnoCableTests as TFofi4PointTentnoCableTests
+from NightlyTests import Fofi4PointTentCableTests as TFofi4PointTentCableTests
+from NightlyTests import MembraneQ4PointLoadTests as TMembraneQ4PointLoadTests
+from NightlyTests import MembraneQ4TrussPointLoadTests as TMembraneQ4TrussPointLoadTests
+# 2Node Element tests
+from NightlyTests import Simple3D2NTrussTest as T3D2NTrussTest
+from NightlyTests import Simple3D2NTrussLinearTest as T3D2NTrussLinearTest
+from NightlyTests import Simple3D2NTrussDynamicTest as T3D2NTrussDynamicTest
+from NightlyTests import Simple3D2NBeamCrTest as T3D2NBeamCrTest
+from NightlyTests import Simple3D2NBeamCrLinearTest as T3D2NBeamCrLinearTest
+from NightlyTests import Simple3D2NBeamCrDynamicTest as T3D2NBeamCrDynamicTest
+from NightlyTests import Simple2D2NBeamCrTest as T2D2NBeamCrTest
+# Shell tests
 from NightlyTests import ShellQ4ThickBendingRollUpTests as TShellQ4ThickBendingRollUpTests
 from NightlyTests import ShellQ4ThickDrillingRollUpTests as TShellQ4ThickDrillingRollUpTests
 from NightlyTests import ShellQ4ThickOrthotropicLaminateLinearStaticTests as TShellQ4ThickOrthotropicLaminateLinearStaticTests
-
 from NightlyTests import ShellT3ThinBendingRollUpTests as TShellT3ThinBendingRollUpTests
 from NightlyTests import ShellT3ThinDrillingRollUpTests as TShellT3ThinDrillingRollUpTests
 from NightlyTests import ShellT3IsotropicScordelisTests as TShellT3IsotropicScordelisTests
 from NightlyTests import ShellT3ThinOrthotropicLaminateLinearStaticTests as TShellT3ThinOrthotropicLaminateLinearStaticTests
-
 from NightlyTests import ShellT3ThickLinearStaticTests as TShellT3ThickLinearStaticTests
 from NightlyTests import ShellT3ThickNonLinearStaticTests as TShellT3ThickNonLinearStaticTests
 from NightlyTests import ShellT3ThickLinearDynamicTests as TShellT3ThickLinearDynamicTests
 from NightlyTests import ShellT3ThickNonLinearDynamicTests as TShellT3ThickNonLinearDynamicTests
 from NightlyTests import ShellT3ThickOrthotropicLaminateLinearStaticTests as TShellT3ThickOrthotropicLaminateLinearStaticTests
-
 from NightlyTests import ShellQ4ThinLinearStaticTests as TShellQ4ThinLinearStaticTests
 from NightlyTests import ShellQ4ThinNonLinearStaticTests as TShellQ4ThinNonLinearStaticTests
 from NightlyTests import ShellQ4ThinLinearDynamicTests as TShellQ4ThinLinearDynamicTests
@@ -121,11 +121,12 @@ from NightlyTests import ShellQ4ThinNonLinearDynamicTests as TShellQ4ThinNonLine
 from NightlyTests import ShellQ4ThinOrthotropicLaminateLinearStaticTests as TShellQ4ThinOrthotropicLaminateLinearStaticTests
 
 # CL tests
-##from NightlyTests import IsotropicDamageSimoJuPSTest    as TIsotropicDamageSimoJuPSTest
+#from NightlyTests import IsotropicDamageSimoJuPSTest    as TIsotropicDamageSimoJuPSTest
 
-## VALIDATION TESTS
+##### VALIDATION TESTS #####
 # SPRISM tests
 #from ValidationTests import SprismPanTests              as TSprismPanTests
+# Pendulus Tests with Solid Elements
 from ValidationTests import PendulusTLTest              as TPendulusTLTest
 from ValidationTests import PendulusULTest              as TPendulusULTest
 
@@ -145,119 +146,122 @@ def AssambleTestSuites():
     suites = KratosUnittest.KratosSuites
 
     # Create a test suit with the selected tests (Small tests):
-    smallSuite = suites['small']
-    # Simple patch tests
-    ## Solids
+    # These tests are executed by the continuous integration tool, so they have to be very fast!
+    # Execution time << 1 sec on a regular PC !!!
+    # If the tests in the smallSuite take too long then merging to master will not be possible!
+    smallSuite = suites['small'] # These tests are executed by the continuous integration tool
+    nightSuite = suites['nightly'] # These tests are executed in the nightly build
+
+    ### Adding the self-contained tests
+    #Solids
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestConstitutiveLaw]))
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestPatchTestSmallStrain]))
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestPatchTestLargeStrain]))
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestQuadraticElements]))
-    ## Shells
+    # Shells
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestPatchTestShells]))
-    smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestPatchTestShellsStressRec]))
-    smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestPatchTestShellsOrthotropic]))
-    ## Trusses
+    nightSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestPatchTestShellsStressRec])) # TODO should be in smallSuite but is too slow
+    nightSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestPatchTestShellsOrthotropic])) # TODO should be in smallSuite but is too slow
+    # Trusses
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestTruss3D2N]))
-    ## Beams
+    # Beams
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestCrBeam3D2N]))
-    smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestCrBeam2D2N]))
+    nightSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestCrBeam2D2N])) # TODO should be in smallSuite but is too slow
     # Test loading conditions
-    smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TestLoadingConditionsPoint]))
-    smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TestLoadingConditionsLine]))
-    smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TestLoadingConditionsSurface]))
-    # Basic moving mesh test
-    smallSuite.addTest(TSimpleMeshMovingTest('test_execution'))
-    # Dynamic basic tests
-    smallSuite.addTest(TDynamicBossakTests('test_execution'))
-    smallSuite.addTest(TDynamicNewmarkTests('test_execution'))
-    # Patch test Small Displacements
-    smallSuite.addTest(TSDTwoDShearQuaPatchTest('test_execution'))
-    smallSuite.addTest(TSDTwoDShearTriPatchTest('test_execution'))
-    smallSuite.addTest(TSDTwoDTensionQuaPatchTest('test_execution'))
-    smallSuite.addTest(TSDTwoDTensionTriPatchTest('test_execution'))
-    smallSuite.addTest(TSDThreeDShearHexaPatchTest('test_execution'))
-    smallSuite.addTest(TSDThreeDShearTetraPatchTest('test_execution'))
-    smallSuite.addTest(TSDThreeDTensionHexaPatchTest('test_execution'))
-    smallSuite.addTest(TSDThreeDTensionTetraPatchTest('test_execution'))
-    # Patch test Total Lagrangian
-    smallSuite.addTest(TTLTwoDShearQuaPatchTest('test_execution'))
-    smallSuite.addTest(TTLTwoDShearTriPatchTest('test_execution'))
-    smallSuite.addTest(TTLTwoDTensionQuaPatchTest('test_execution'))
-    smallSuite.addTest(TTLTwoDTensionTriPatchTest('test_execution'))
-    smallSuite.addTest(TTLThreeDShearHexaPatchTest('test_execution'))
-    smallSuite.addTest(TTLThreeDShearTetraPatchTest('test_execution'))
-    smallSuite.addTest(TTLThreeDTensionHexaPatchTest('test_execution'))
-    smallSuite.addTest(TTLThreeDTensionTetraPatchTest('test_execution'))
-    # Patch test Updated Lagrangian
-    smallSuite.addTest(TULTwoDShearQuaPatchTest('test_execution'))
-    smallSuite.addTest(TULTwoDShearTriPatchTest('test_execution'))
-    smallSuite.addTest(TULTwoDTensionQuaPatchTest('test_execution'))
-    smallSuite.addTest(TULTwoDTensionTriPatchTest('test_execution'))
-    smallSuite.addTest(TULThreeDShearHexaPatchTest('test_execution'))
-    smallSuite.addTest(TULThreeDShearTetraPatchTest('test_execution'))
-    smallSuite.addTest(TULThreeDTensionHexaPatchTest('test_execution'))
-    smallSuite.addTest(TULThreeDTensionTetraPatchTest('test_execution'))
-    # SPRISM tests
-    smallSuite.addTest(TSprismMembranePatchTests('test_execution'))
-    smallSuite.addTest(TSprismBendingPatchTests('test_execution'))
-    # Membrane tests
-    smallSuite.addTest(TFofi4PointTentnoCableTests('test_execution'))
-    smallSuite.addTest(TFofi4PointTentCableTests('test_execution'))
-    smallSuite.addTest(TMembraneQ4PointLoadTests('test_execution'))
-    smallSuite.addTest(TMembraneQ4TrussPointLoadTests('test_execution'))
-    # 2Node Element tests    
-    smallSuite.addTest(T3D2NTrussDynamicTest('test_execution'))
-    smallSuite.addTest(T3D2NTrussLinearTest('test_execution'))
-    smallSuite.addTest(T3D2NTrussTest('test_execution'))
-    smallSuite.addTest(T3D2NBeamCrTest('test_execution'))
-    smallSuite.addTest(T3D2NBeamCrLinearTest('test_execution'))
-    smallSuite.addTest(T3D2NBeamCrDynamicTest('test_execution'))
+    smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestLoadingConditionsPoint]))
+    smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestLoadingConditionsLine]))
+    smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestLoadingConditionsSurface]))
     # Nodal damping test
-    smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TNodalDampingTests]))
+    nightSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TNodalDampingTests])) # TODO should be in smallSuite but is too slow
+    # Multipoint constraint tests
+    smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TTestMultipointConstraints]))
+    
+    ### Adding Small Tests
+    # Dynamic basic tests (leave these in the smallSuite to have the Exection script tested)
+    smallSuite.addTest(TDynamicBossakTests('test_execution')) 
+    smallSuite.addTest(TDynamicNewmarkTests('test_execution'))
 
-    if (missing_external_dependencies == False):
-        if (hasattr(KratosMultiphysics.ExternalSolversApplication,
-                    "FEASTSolver")):
-            # Eigenvalues tests
-            smallSuite.addTest(TEigenQ4Thick2x2PlateTests('test_execution'))
-            smallSuite.addTest(TEigen3D3NThinCircleTests('test_execution'))
-            smallSuite.addTest(TEigenTL3D8NCubeTests('test_execution'))
-            # Element damping test
-            smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TSpringDamperElementTests]))
-            # Harmonic analysis test
-            smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([THarmonicAnalysisTests]))
-        else:
-            print(
-                "FEASTSolver solver is not included in the compilation of the External Solvers Application"
-            )
-
-
-    # Create a test suit with the selected tests plus all small tests
-    nightSuite = suites['nightly']
-    nightSuite.addTests(smallSuite)
+    ### Adding Nightly Tests
+    # Basic moving mesh test
+    nightSuite.addTest(TSimpleMeshMovingTest('test_execution'))
+    # Patch test Small Displacements
+    nightSuite.addTest(TSDTwoDShearQuaPatchTest('test_execution'))
+    nightSuite.addTest(TSDTwoDShearTriPatchTest('test_execution'))
+    nightSuite.addTest(TSDTwoDTensionQuaPatchTest('test_execution'))
+    nightSuite.addTest(TSDTwoDTensionTriPatchTest('test_execution'))
+    nightSuite.addTest(TSDThreeDShearHexaPatchTest('test_execution'))
+    nightSuite.addTest(TSDThreeDShearTetraPatchTest('test_execution'))
+    nightSuite.addTest(TSDThreeDTensionHexaPatchTest('test_execution'))
+    nightSuite.addTest(TSDThreeDTensionTetraPatchTest('test_execution'))
+    # Patch test Total Lagrangian
+    nightSuite.addTest(TTLTwoDShearQuaPatchTest('test_execution'))
+    nightSuite.addTest(TTLTwoDShearTriPatchTest('test_execution'))
+    nightSuite.addTest(TTLTwoDTensionQuaPatchTest('test_execution'))
+    nightSuite.addTest(TTLTwoDTensionTriPatchTest('test_execution'))
+    nightSuite.addTest(TTLThreeDShearHexaPatchTest('test_execution'))
+    nightSuite.addTest(TTLThreeDShearTetraPatchTest('test_execution'))
+    nightSuite.addTest(TTLThreeDTensionHexaPatchTest('test_execution'))
+    nightSuite.addTest(TTLThreeDTensionTetraPatchTest('test_execution'))
+    # Patch test Updated Lagrangian
+    nightSuite.addTest(TULTwoDShearQuaPatchTest('test_execution'))
+    nightSuite.addTest(TULTwoDShearTriPatchTest('test_execution'))
+    nightSuite.addTest(TULTwoDTensionQuaPatchTest('test_execution'))
+    nightSuite.addTest(TULTwoDTensionTriPatchTest('test_execution'))
+    nightSuite.addTest(TULThreeDShearHexaPatchTest('test_execution'))
+    nightSuite.addTest(TULThreeDShearTetraPatchTest('test_execution'))
+    nightSuite.addTest(TULThreeDTensionHexaPatchTest('test_execution'))
+    nightSuite.addTest(TULThreeDTensionTetraPatchTest('test_execution'))
+    # SPRISM tests
+    nightSuite.addTest(TSprismMembranePatchTests('test_execution'))
+    nightSuite.addTest(TSprismBendingPatchTests('test_execution'))
+    # Membrane tests
+    nightSuite.addTest(TFofi4PointTentnoCableTests('test_execution'))
+    nightSuite.addTest(TFofi4PointTentCableTests('test_execution'))
+    nightSuite.addTest(TMembraneQ4PointLoadTests('test_execution'))
+    nightSuite.addTest(TMembraneQ4TrussPointLoadTests('test_execution'))
+    # 2Node Element tests    
+    nightSuite.addTest(T3D2NTrussDynamicTest('test_execution'))
+    nightSuite.addTest(T3D2NTrussLinearTest('test_execution'))
+    nightSuite.addTest(T3D2NTrussTest('test_execution'))
+    nightSuite.addTest(T3D2NBeamCrTest('test_execution'))
+    nightSuite.addTest(T3D2NBeamCrLinearTest('test_execution'))
+    nightSuite.addTest(T3D2NBeamCrDynamicTest('test_execution'))
     # Shell tests
     nightSuite.addTest(TShellQ4ThickBendingRollUpTests('test_execution'))
     nightSuite.addTest(TShellQ4ThickDrillingRollUpTests('test_execution'))
     nightSuite.addTest(TShellQ4ThickOrthotropicLaminateLinearStaticTests('test_execution'))
-
     nightSuite.addTest(TShellT3ThinBendingRollUpTests('test_execution'))
     nightSuite.addTest(TShellT3ThinOrthotropicLaminateLinearStaticTests('test_execution'))
-
     nightSuite.addTest(TShellT3ThickLinearStaticTests('test_execution'))
     nightSuite.addTest(TShellT3ThickNonLinearStaticTests('test_execution'))
     nightSuite.addTest(TShellT3ThickLinearDynamicTests('test_execution'))
     nightSuite.addTest(TShellT3ThickNonLinearDynamicTests('test_execution'))
     nightSuite.addTest(TShellT3ThickOrthotropicLaminateLinearStaticTests('test_execution'))
-
     nightSuite.addTest(TShellQ4ThinLinearStaticTests('test_execution'))
     nightSuite.addTest(TShellQ4ThinNonLinearStaticTests('test_execution'))
     nightSuite.addTest(TShellQ4ThinLinearDynamicTests('test_execution'))
     nightSuite.addTest(TShellQ4ThinNonLinearDynamicTests('test_execution'))
     nightSuite.addTest(TShellQ4ThinOrthotropicLaminateLinearStaticTests('test_execution'))
-	
+    
     # CL tests
-    ##nightSuite.addTest(TIsotropicDamageSimoJuPSTest('test_execution')) # FIXME: Needs get up to date
+    #nightSuite.addTest(TIsotropicDamageSimoJuPSTest('test_execution')) # FIXME: Needs get up to date
 
+    if (missing_external_dependencies == False):
+        if (hasattr(KratosMultiphysics.ExternalSolversApplication, "FEASTSolver")):
+            # Eigenvalues tests
+            smallSuite.addTest(TEigenQ4Thick2x2PlateTests('test_execution'))
+            smallSuite.addTest(TEigenTL3D8NCubeTests('test_execution'))
+            nightSuite.addTest(TEigen3D3NThinCircleTests('test_execution'))
+            # Harmonic analysis test
+            smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([THarmonicAnalysisTests]))
+            # Element damping test
+            nightSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([TSpringDamperElementTests])) # TODO should be in smallSuite but is too slow
+        else:
+            print("FEASTSolver solver is not included in the compilation of the External Solvers Application")
+
+    nightSuite.addTests(smallSuite)
+
+    ### Adding Validation Tests
     # For very long tests that should not be in nighly and you can use to validate 
     validationSuite = suites['validation']
     # SPRISM tests
@@ -266,7 +270,7 @@ def AssambleTestSuites():
     validationSuite.addTest(TPendulusULTest('test_execution'))
     validationSuite.addTest(TShellT3ThinDrillingRollUpTests('test_execution'))
     validationSuite.addTest(TShellT3IsotropicScordelisTests('test_execution'))
-    validationSuite.addTest(T2D2NBeamCrTest('test_execution'))
+    validationSuite.addTest(T2D2NBeamCrTest('test_execution')) # TODO should be in nightSuite but is too slow
     
     # Create a test suit that contains all the tests:
     allSuite = suites['all']
